"rebase" classes on spring-context:5.1.4.RELEASE
The code for most of our classes is copied from spring-context with a
small amount changes in order to support using
com.ibm.icu.text.MessageFormat in place of java.text.MessageFormat.

rebasing the copied code against a newer version of Spring allows us to
delete the CacheEntry and Pair util classes and pulls in any
improvements or bug fixes that have been made to the code by the Spring
developers.