"rebase" code against spring-context 5.2.5.RELEASE
The code for most of our classes is copied from spring-context with a
small amount changes in order to support using
com.ibm.icu.text.MessageFormat in place of java.text.MessageFormat.

This PR keeps the copied code up to date with the latest version
(5.2.5.RELEASE) of spring-context. This PR also updates the other libs
to their latest versions.